+++ apps/qsdm-hive/qsdm-hive-main/src/renderer/features/notifications/Banners/UpdateAvailableNotification.tsx
@@ -2,8 +2,6 @@
 
 import React, { useEffect } from 'react';
 
-import { downloadAppUpdate } from 'renderer/services';
-
 import { BackButtonSlotType, NotificationType } from '../types';
 import { useNotificationActions } from '../useNotificationStore';
 
@@ -18,12 +16,10 @@ export function UpdateAvailableNotification({
 }) {
   const { markAsRead } = useNotificationActions();
 
-  const [isDownloading, setIsDownloading] = React.useState(false);
   const [isDownloaded, setIsDownloaded] = React.useState(false);
 
   useEffect(() => {
     const destroy = window.main.onAppDownloaded(() => {
-      setIsDownloading(() => false);
       setIsDownloaded(() => true);
 
       setTimeout(() => {
@@ -34,43 +30,24 @@ export function UpdateAvailableNotification({
     return () => {
       destroy();
     };
-  }, [isDownloaded, markAsRead, notification.id]);
-
-  const handleUpdateDownload = () => {
-    setIsDownloading(true);
-    downloadAppUpdate().catch((err) => {
-      console.error(err);
-    });
-  };
+  }, [markAsRead, notification.id]);
 
   const getContent = () => {
-    if (isDownloading) {
-      return 'New version is downloading...';
-    }
-
     if (isDownloaded) {
-      return 'New version is ready to install!';
+      return 'Ready to restart';
     }
 
-    return (
-      <button
-        className="text-white bg-transparent text-white font-semibold w-max px-4 py-2 rounded-[4px] border-white border-2  hover:bg-[rgb(243,243,247)]/[0.1] hover:shadow-xl transition-all duration-300 ease-in-out"
-        onClick={handleUpdateDownload}
-      >
-        Update Now
-      </button>
-    );
+    return 'Downloading...';
   };
-
-  const showBannerMainContent = !isDownloading && !isDownloaded;
 
   return (
     <NotificationDisplayBanner
       notification={notification}
       messageSlot={
         <div className="">
-          {showBannerMainContent &&
-            'A new version of the node is ready for you!'}
+          {isDownloaded
+            ? 'The required QSDM Hive update is verified.'
+            : 'A required QSDM Hive update is downloading.'}
         </div>
       }
       actionButtonSlot={getContent()}